api/requirements.txt
tensorflow==2.19.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.19.0
matplotlib
numpy